Moved classes `Player` and `Network` to new file
