+++ website/__init__.py
@@ -4,6 +4,8 @@ This code is run once at the start of the game
 
 import eventlet
 
+from .database.player import Player
+
 eventlet.monkey_patch(thread=True, time=True)
 
 from flask import Flask  # noqa: E402
@@ -71,7 +73,7 @@ def create_app(run_init_test_players, rm_instance):
     app.register_blueprint(http, url_prefix="/")
     app.register_blueprint(ws, url_prefix="/")
 
-    from .database.database import Hex, Player
+    from .database.database import Hex
 
     # initialize database :
     with app.app_context():
