+++ website/api/ws.py
@@ -9,8 +9,9 @@ from werkzeug.security import check_password_hash
 from simple_websocket import ConnectionClosed
 
 from website import utils
+from ..database.player import Network, Player
 
-from ..database.database import Hex, Player, Network
+from ..database.database import Hex
 
 ws = Blueprint("rest_api", __name__)
