+++ website/auth.py
@@ -12,7 +12,7 @@ from flask import (
     g,
 )
 from flask import current_app
-from .database.database import Player
+from .database.player import Player
 from werkzeug.security import generate_password_hash, check_password_hash
 from . import db
 from flask_login import login_user, login_required, logout_user, current_user
