+++ website/utils.py
@@ -16,9 +16,8 @@ import json
 
 
 from website.api import ws
+from .database.player import Network, Player
 from .database.database import (
-    Player,
-    Network,
     Resource_on_sale,
     Shipment,
     Chat,
